add catboost framework support to ml-wrappers
