+++ requirements-dev.txt
@@ -1,5 +1,6 @@
 lightgbm
 xgboost
+catboost
 tensorflow
 shap
 transformers<4.20.0
